Set router
