+++ 02-express-tutorial/app.js
@@ -1,79 +1,19 @@
 const express = require('express');
+const { people } = require('./data');
 const app = express();
 
-let { people } = require('./data');
-
-app.post('/login', (req, res) => {
-  const { name } = req.body;
-  if (name) {
-    return res.status(200).send(`Welcome ${name}`);
-  }
-  res.status(401).send('Please Provide Credentials');
-});
+const people = require('./routes/people');
+const auth = require('./routes/auth');
 
 // static assets
 app.use(express.static('./methods-public'));
 // parse form data
 app.use(express.urlencoded({ extended: false }));
-
 // parse json
 app.use(express.json());
-app.get('/api/people', (req, res) => {
-  res.status(200).json({ success: true, data: people });
-});
 
-app.post('/api/people', (req, res) => {
-  const { name } = req.body;
-  if (!name) {
-    return res
-      .status(400)
-      .json({ success: false, msg: 'please provide name value' });
-  }
-  res.status(201).send({ Success: true, person: name });
-});
-
-app.post('/api/postman/people', (req, res) => {
-  const { name } = req.body;
-  if (!name) {
-    return res
-      .status(400)
-      .json({ success: false, msg: 'please provide name value' });
-  }
-  res.status(201).send({ success: true, data: [...people, name] });
-});
-
-app.put('/api/people/:id', (req, res) => {
-  const { id } = req.params;
-  const { name } = req.body;
-
-  const person = people.find((person) => person.id === Number(id));
-
-  if (!person) {
-    return res
-      .status(404)
-      .json({ success: false, msg: `no person with id ${id}` });
-  }
-  const newPeople = people.map((person) => {
-    if (person.id === Number(id)) {
-      person.name = name;
-    }
-    return person;
-  });
-  res.status(200).json({ success: true, data: newPeople });
-});
-
-app.delete('/api/people/:id', (req, req) => {
-  const person = people.find((person) => person.id === Number(req.params.id));
-  if (!person) {
-    return res
-      .status(404)
-      .json({ success: false, msg: `no person with id ${id}` });
-  }
-  const newPeople = people.filter(
-    (person) => person.id !== Number(req.params.id),
-  );
-  return res.status(200).json({ success: true, data: newPeople });
-});
+app.use('api/people', people);
+app.use('/login', auth);
 
 app.listen(5000, () => {
   console.log('Server is running on port 5000...');
