+++ 02-express-tutorial/routes/auth.js
@@ -1,0 +1,12 @@
+const express = require('express');
+const router = express.Router();
+
+router.post('/', (req, res) => {
+  const { name } = req.body;
+  if (name) {
+    return res.status(200).send(`Welcome ${name}`);
+  }
+  res.status(401).send('Please Provide Credentials');
+});
+
+module.exports = router;
